+++ src/apyfloatarray_wrapper.cc
@@ -347,11 +347,11 @@ void bind_float_array(nb::module_& m)
                 Floating point values to initialize from. The tensor shape will be taken
                 from the sequence shape.
             exp_bits : int
-                Number of exponent bits in the created fixed-point tensor
+                Number of exponent bits in the created floating-point tensor
             man_bits : int
-                Number of mantissa bits in the created fixed-point tensor
+                Number of mantissa bits in the created floating-point tensor
             bias : int, optional
-                Bias in the created fixed-point tensor
+                Bias in the created floating-point tensor
 
             Returns
             -------
@@ -393,11 +393,11 @@ void bind_float_array(nb::module_& m)
             ndarray : ndarray
                 Values to initialize from. The tensor shape will be taken from the ndarray shape.
             exp_bits : int
-                Number of exponent bits in the created fixed-point tensor
+                Number of exponent bits in the created floating-point tensor
             man_bits : int
-                Number of mantissa bits in the created fixed-point tensor
+                Number of mantissa bits in the created floating-point tensor
             bias : int, optional
-                Bias in the created fixed-point tensor
+                Bias in the created floating-point tensor
 
             Returns
             -------