+++ CHANGELOG.md
@@ -30,6 +30,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Fix bug in `APyFixedArray.__truediv__()` where program crashes on long-limb zero
   denominators.
 - Fix bug where single-dimensional tuple representations missed out on a comma.
+- Fix bug in comparison between floating-point infinity and fixed-point numbers.
 
 ### Removed